Packed Changed to use Nimble Methods
Now requires Nimble.
The need to compile code will make this version slower than the non-Nimble version for small runs but will be much quicker for larger amounts of samples.
The MLE isn't implemented in this so its been removed and so has the code for the prior probability, random walk, independence sampler etc. since these are already included in Nimble.